Clean up docstrings, rename GroupMask file
