+++ packages/sdk/src/types/runtime/index.ts
@@ -18,4 +18,4 @@ export * from './assets';
 export * from './physics';
 export * from './lookAt';
 export * from './appearance';
-export * from './userGroupCollection';
+export * from './groupMask';
